// src/models/npcs/species/humanoids/religions/traits/types.ts
import { Trait } from '../../../../../utilities/traits/types'
import type { Religion } from '../types'

export type religion__tags =
  | 'ancestor worship'
  | 'antinomian clergy'
  | 'aristocratic'
  | 'asceticism'
  | 'astrology'
  | 'caste structure'
  | 'charity'
  | 'decadence'
  | 'dire sacrifices'
  | 'economic role'
  | 'esotericism'
  | 'gnosticism'
  | 'harmonious'
  | 'holy grounds'
  | 'idol worship'
  | 'legalism'
  | 'localized'
  | 'materially luxuriant'
  | 'militant'
  | 'missionary zeal'
  | 'naturalism'
  | 'pacifist'
  | 'quietist'
  | 'reincarnation'
  | 'ritualism'
  | 'simony'
  | 'supremacy'
  | 'underclass'

export type ReligionTrait = Trait<religion__tags, { entity: Religion }>